requirements.txt
matplotlib==3.6.1
librosa==0.8.0
tqdm==4.65.0
pandas==2.0.0
numba==0.53.1
numpy==1.21.0
scipy==1.6.2
tensorboard==2.11.0
PyYAML==6.0
tensorboardX==2.1
pyloudnorm==0.1.0
setuptools==61.2.0
g2p_en==2.1.0
resemblyzer==0.1.1.dev0
webrtcvad==2.0.10
pycwt==0.3.0a22
praat-parselmouth==0.3.3
einops==0.3.0
chardet==4.0.0
h5py==3.1.0
transformers==4.28.1
scikit-image==0.16.2
torch==1.9.0+cu111
torchaudio==0.9.0
textgrid===1.5
pillow==8.0.1
imageio==2.9.0
packaging==23.1